don't download mdx if it already exists

+++ make/ci.mk
@@ -42,7 +42,7 @@ checker.deadlink:
 .PHONY: checker.quickstart
 checker.quickstart:
 	@echo "===========> Checking QuickStart Doc"
-	curl -o ${ROOT_DIR}/mdx https://raw.githubusercontent.com/seeflood/mdx/main/mdx
+	if ! test -e mdx; then curl -o ${ROOT_DIR}/mdx https://raw.githubusercontent.com/seeflood/mdx/main/mdx ;fi
 	mv ${ROOT_DIR}/mdx ${SCRIPT_DIR}
 	chmod +x  ${SCRIPT_DIR}/mdx
 	sh ${SCRIPT_DIR}/test-quickstart.sh
@@ -50,7 +50,7 @@ checker.quickstart:
 .PHONY: checker.quickstart.17
 checker.quickstart.17:
 	@echo "===========> Checking QuickStart Doc"
-	curl -o ${ROOT_DIR}/mdx https://raw.githubusercontent.com/seeflood/mdx/main/mdx
+	if ! test -e mdx; then curl -o ${ROOT_DIR}/mdx https://raw.githubusercontent.com/seeflood/mdx/main/mdx ;fi
 	mv ${ROOT_DIR}/mdx ${SCRIPT_DIR}
 	chmod +x  ${SCRIPT_DIR}/mdx
 	sh ${SCRIPT_DIR}/test-quickstart-17.sh
